引擎/requirements.txt
# 低端GPU优化渲染引擎依赖列表

# 核心依赖
numpy>=1.21.0  # 优化的数值计算，对低端GPU非常重要
pyglet>=2.0.0  # 窗口管理和输入处理
moderngl>=5.6.4  # 现代OpenGL包装器，支持着色器和纹理压缩
PyOpenGL>=3.1.5  # 兼容性OpenGL支持
Pillow>=8.3.0  # 图像处理和纹理优化
PyGLM>=2.4.0  # 高性能GL数学库，优化过的矩阵/向量运算
pyrr>=0.10.3  # 额外的3D数学支持

# 性能关键优化依赖
# numba>=0.55.0  # Python JIT编译，为核心渲染路径提供接近C的性能 - 暂时注释，不支持Python 3.14
cython>=0.29.24  # C扩展生成，用于最关键的渲染函数
scipy>=1.7.0  # 科学计算，用于LOD生成和优化算法

# 硬件检测与优化
py-cpuinfo>=8.0.0  # CPU信息检测，用于平台特定优化
psutil>=5.8.0  # 系统资源监控，用于实时性能调整
py-spy>=0.3.10  # 性能分析工具，用于识别瓶颈

# 开发和测试
pytest>=6.2.5  # 测试框架，确保优化不会破坏功能
black>=22.0.0  # 代码格式化
flake8>=4.0.0  # 代码检查

# 调试和可视化（开发环境使用）
matplotlib>=3.4.0  # 用于性能分析和可视化

# 图像和视频处理
imageio>=2.9.0  # 图像读取和处理，支持各种格式

# 平台特定优化（可选）
# pycuda>=2021.1  # NVIDIA GPU计算优化（仅NVIDIA GPU）
# pyopencl>=2021.2.6  # 通用GPU计算优化

# Blender集成注意事项
# bpy>=2.93.0  # Blender Python API，需要单独安装
# 请参考Blender官方文档获取bpy的安装方法，不在PyPI上直接提供

# 安装说明：
# pip install -r requirements.txt
# 对于最佳性能，确保所有可选优化库都已安装
# 对于低端GPU测试，建议使用虚拟环境以避免依赖冲突